moved PyMouseEvent to events module
