+++ pymouse/__init__.py
@@ -23,14 +23,25 @@ See http://github.com/pepijndevos/PyMouse for more information.
 import sys
 
 if sys.platform.startswith('java'):
-    from java_ import PyMouse
+    from pymouse.java_ import PyMouse
 
 elif sys.platform == 'darwin':
-    from mac import PyMouse, PyMouseEvent
+    from pymouse.mac import PyMouse
+    try:
+        from pymouse.events.mac import PyMouseEvent
+    except ImportError:
+        from pymouse.base import PyMouseEventMeta as PyMouseEvent
 
 elif sys.platform == 'win32':
-    from windows import PyMouse, PyMouseEvent
+    from pymouse.windows import PyMouse
+    try:
+        from pymouse.events.windows import PyMouseEvent
+    except ImportError:
+        from pymouse.base import PyMouseEventMeta as PyMouseEvent
 
 else:
-    from unix import PyMouse, PyMouseEvent
-
+    from pymouse.unix import PyMouse
+    try:
+        from pymouse.events.unix import PyMouseEvent
+    except ImportError:
+        from pymouse.base import PyMouseEventMeta as PyMouseEvent
