+++ pymouse/events/mac.py
@@ -1,0 +1,57 @@
+# -*- coding: iso-8859-1 -*-
+
+#   Copyright 2010 Pepijn de Vos
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+
+from Quartz import *
+from AppKit import NSEvent
+from pymouse.base import PyMouseEventMeta
+
+class PyMouseEvent(PyMouseEventMeta):
+    def run(self):
+        tap = CGEventTapCreate(
+            kCGSessionEventTap,
+            kCGHeadInsertEventTap,
+            kCGEventTapOptionDefault,
+            CGEventMaskBit(kCGEventMouseMoved) |
+            CGEventMaskBit(kCGEventLeftMouseDown) |
+            CGEventMaskBit(kCGEventLeftMouseUp) |
+            CGEventMaskBit(kCGEventRightMouseDown) |
+            CGEventMaskBit(kCGEventRightMouseUp) |
+            CGEventMaskBit(kCGEventOtherMouseDown) |
+            CGEventMaskBit(kCGEventOtherMouseUp),
+            self.handler,
+            None)
+
+        loopsource = CFMachPortCreateRunLoopSource(None, tap, 0)
+        loop = CFRunLoopGetCurrent()
+        CFRunLoopAddSource(loop, loopsource, kCFRunLoopDefaultMode)
+        CGEventTapEnable(tap, True)
+
+        while self.state:
+            CFRunLoopRunInMode(kCFRunLoopDefaultMode, 5, False)
+
+    def handler(self, proxy, type, event, refcon):
+        (x, y) = CGEventGetLocation(event)
+        if type in pressID:
+            self.click(x, y, pressID.index(type), True)
+        elif type in releaseID:
+            self.click(x, y, releaseID.index(type), False)
+        else:
+            self.move(x, y)
+
+        if self.capture:
+            CGEventSetType(event, kCGEventNull)
+
+        return event
