+++ pymouse/events/windows.py
@@ -1,0 +1,58 @@
+# -*- coding: iso-8859-1 -*-
+
+#   Copyright 2010 Pepijn de Vos
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+
+from pymouse.base import PyMouseEventMeta
+import pythoncom, pyHook
+from time import sleep
+
+class PyMouseEvent(PyMouseEventMeta):
+    def __init__(self):
+        PyMouseEventMeta.__init__(self)
+        self.hm = pyHook.HookManager()
+
+    def run(self):
+        self.hm.MouseAllButtons = self._click
+        self.hm.MouseMove = self._move
+        self.hm.HookMouse()
+        while self.state:
+            sleep(0.01)
+            pythoncom.PumpWaitingMessages()
+
+    def stop(self):
+        self.hm.UnhookMouse()
+        self.state = False
+
+    def _click(self, event):
+        x,y = event.Position
+
+        if event.Message == pyHook.HookConstants.WM_LBUTTONDOWN:
+            self.click(x, y, 1, True)
+        elif event.Message == pyHook.HookConstants.WM_LBUTTONUP:
+            self.click(x, y, 1, False)
+        elif event.Message == pyHook.HookConstants.WM_RBUTTONDOWN:
+            self.click(x, y, 2, True)
+        elif event.Message == pyHook.HookConstants.WM_RBUTTONUP:
+            self.click(x, y, 2, False)
+        elif event.Message == pyHook.HookConstants.WM_MBUTTONDOWN:
+            self.click(x, y, 3, True)
+        elif event.Message == pyHook.HookConstants.WM_MBUTTONUP:
+            self.click(x, y, 3, False)
+        return not self.capture
+
+    def _move(self, event):
+        x,y = event.Position
+        self.move(x, y)
+        return not self.captureMove
