+++ pymouse/java_.py
@@ -17,7 +17,7 @@
 from java.awt import Robot, Toolkit
 from java.awt.event import InputEvent
 from java.awt.MouseInfo import getPointerInfo
-from base import PyMouseMeta
+from pymouse.base import PyMouseMeta
 
 r = Robot()
