+++ pymouse/windows.py
@@ -16,22 +16,20 @@
 
 from ctypes import *
 import win32api,win32con
-from base import PyMouseMeta, PyMouseEventMeta
-import pythoncom, pyHook
-from time import sleep
+from pymouse.base import PyMouseMeta, PyMouseEventMeta
 
 class POINT(Structure):
     _fields_ = [("x", c_ulong),
                 ("y", c_ulong)]
 
 class PyMouse(PyMouseMeta):
-    """MOUSEEVENTF_(button and action) constants 
+    """MOUSEEVENTF_(button and action) constants
     are defined at win32con, buttonAction is that value"""
     def press(self, x, y, button = 1):
         buttonAction = 2**((2*button)-1)
         self.move(x,y)
         win32api.mouse_event(buttonAction, x, y)
-     
+
     def release(self, x, y, button = 1):
         buttonAction = 2**((2*button))
         self.move(x,y)
@@ -49,42 +47,3 @@ class PyMouse(PyMouseMeta):
         width = windll.user32.GetSystemMetrics(0)
         height = windll.user32.GetSystemMetrics(1)
         return width, height
-
-class PyMouseEvent(PyMouseEventMeta):
-    def __init__(self):
-        PyMouseEventMeta.__init__(self)
-        self.hm = pyHook.HookManager()
-
-    def run(self):
-        self.hm.MouseAllButtons = self._click
-        self.hm.MouseMove = self._move
-        self.hm.HookMouse()
-        while self.state:
-            sleep(0.01)
-            pythoncom.PumpWaitingMessages()
-
-    def stop(self):
-        self.hm.UnhookMouse()
-        self.state = False
-
-    def _click(self, event):
-        x,y = event.Position
-
-        if event.Message == pyHook.HookConstants.WM_LBUTTONDOWN:
-            self.click(x, y, 1, True)
-        elif event.Message == pyHook.HookConstants.WM_LBUTTONUP:
-            self.click(x, y, 1, False)
-        elif event.Message == pyHook.HookConstants.WM_RBUTTONDOWN:
-            self.click(x, y, 2, True)
-        elif event.Message == pyHook.HookConstants.WM_RBUTTONUP:
-            self.click(x, y, 2, False)
-        elif event.Message == pyHook.HookConstants.WM_MBUTTONDOWN:
-            self.click(x, y, 3, True)
-        elif event.Message == pyHook.HookConstants.WM_MBUTTONUP:
-            self.click(x, y, 3, False)
-        return not self.capture
-
-    def _move(self, event):
-        x,y = event.Position
-        self.move(x, y)
-        return not self.captureMove
